Refactor: move handlers and tasks
